Changed flow of working with directories for monsoon.
I made changes to create a directory prior to using cd .. as this puts students into an unfamiliar place on Monsoon that may cause confusion. Now they first create the directory, use `ls` to see the directory, `cd` into that directory, `ls` again, and then `cd ..` out of that directory.